@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ConferenceMapComponent } from './conference-map.component';
 
@@ -6,7 +6,7 @@ describe('ConferenceMapComponent', () => {
   let component: ConferenceMapComponent;
   let fixture: ComponentFixture<ConferenceMapComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ConferenceMapComponent ]
     })
