@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ConferenceStepsComponent } from './conference-steps.component';
 
@@ -6,7 +6,7 @@ describe('ConferenceStepsComponent', () => {
   let component: ConferenceStepsComponent;
   let fixture: ComponentFixture<ConferenceStepsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ConferenceStepsComponent ]
     })
