@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { HowItWorksSlimComponent } from './how-it-works-slim.component';
 
@@ -6,7 +6,7 @@ describe('HowItWorksSlimComponent', () => {
   let component: HowItWorksSlimComponent;
   let fixture: ComponentFixture<HowItWorksSlimComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ HowItWorksSlimComponent ]
     })
