@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ParticipationsComponent } from './participations.component';
 
@@ -6,7 +6,7 @@ describe('ParticipationsComponent', () => {
   let component: ParticipationsComponent;
   let fixture: ComponentFixture<ParticipationsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ParticipationsComponent ]
     })
