@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ProposalsComponent } from './proposals.component';
 
@@ -6,7 +6,7 @@ describe('ProposalsComponent', () => {
   let component: ProposalsComponent;
   let fixture: ComponentFixture<ProposalsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ProposalsComponent ]
     })
