@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { StrategicAreaComponent } from './strategic-area.component';
 
@@ -6,7 +6,7 @@ describe('StrategicAreaComponent', () => {
   let component: StrategicAreaComponent;
   let fixture: ComponentFixture<StrategicAreaComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ StrategicAreaComponent ]
     })
