@@ -1,5 +1,5 @@
 export interface ISelfDeclarations {
   conference: number;
   locality: number;
-  receiveInformational?: Boolean;
+  receiveInformational?: boolean;
 }
